cvserver/requirements.txt
cvlib~=0.2.7
numpy~=1.22.3
Pillow~=9.0.1